WavExtractor: Skip to data start position if position reset to 0
PiperOrigin-RevId: 260970865

+++ library/core/src/main/java/com/google/android/exoplayer2/extractor/wav/WavHeaderReader.java
@@ -139,7 +139,7 @@ import java.io.IOException;
     // Skip past the "data" header.
     input.skipFully(ChunkHeader.SIZE_IN_BYTES);
 
-    wavHeader.setDataBounds(input.getPosition(), chunkHeader.size);
+    wavHeader.setDataBounds((int) input.getPosition(), chunkHeader.size);
   }
 
   private WavHeaderReader() {
